chore: 🗂 move intlFormatNumber to new format/ folder
