+++ src/core/__tests__/intlFormatNumber.test.ts
@@ -1,6 +1,6 @@
 import { CurrencyEnum } from '~/generated/graphql'
 
-import { intlFormatNumber, getCurrencySymbol } from '../intlFormatNumber'
+import { intlFormatNumber, getCurrencySymbol } from '../formats/intlFormatNumber'
 
 describe('Currency tools', () => {
   describe('intlFormatNumber()', () => {
